Make has_plan_files and at_project_root actual sugar instead of aliases.
This means that plan_files and project_root now require their
arguments and will error if they are not given or are not valid.

Additionally, files passed into project_root, plan_files, plan_file,
and see_also are expanded to absolute paths before being passed up
to Configuration to prevent issues when they are actually loaded.